+++ cob-fianance-app/src/components/layout/BankProgress.js
