+++ cob-fianance-app/src/components/layout/SearchBar.js
