+++ cob-fianance-app/src/components/pages/instructor/SnapshotGroup.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import ProfitProgress from '../../Layout/ProfitProgress';
-import ExpenseProgress from '../../Layout/ExpenseProgress';
-import ExpenseTable from '../../Layout/ExpenseTable';
-import BankProgress from '../../Layout/BankProgress';
+import ProfitProgress from '../../layout/ProfitProgress';
+import ExpenseProgress from '../../layout/ExpenseProgress';
+import ExpenseTable from '../../layout/ExpenseTable';
+import BankProgress from '../../layout/BankProgress';
 import Button from 'react-bootstrap/Button';
 import { API_PATH } from '../../Config';
