+++ cob-fianance-app/src/components/pages/student/Dashboard.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react'
-import ProfitProgress from '../../Layout/ProfitProgress';
-import ExpenseProgress from '../../Layout/ExpenseProgress';
-import BankProgress from '../../Layout/BankProgress';
-import AddTransactionDialogButton from '../../Layout/AddTransactionDialogButton';
-import AddExpenseDialogButton from '../../Layout/AddExpenseDialogButton';
+import ProfitProgress from '../../layout/ProfitProgress';
+import ExpenseProgress from '../../layout/ExpenseProgress';
+import BankProgress from '../../layout/BankProgress';
+import AddTransactionDialogButton from '../../layout/AddTransactionDialogButton';
+import AddExpenseDialogButton from '../../layout/AddExpenseDialogButton';
 import { API_PATH } from '../../Config';
 
 export class Dashboard extends Component {
